refactor(gui): split graphLayout by responsibility
- graphLayout.ts (824 lines) -> entry shell (88) re-exporting stable API
- threeLaneLayout / simpleEgoLayout / claimCoverage / graphLayoutShared /
  graphLayoutTypes extracted by responsibility, all files under gate limit
- reword one comment to satisfy renderer implementation-contracts lint
- gates: check:local fast tier green, test:gui 48/48
